feat: implement robust error handling and comprehensive code quality sweep across workspace

+++ chopin-pg/examples/bench_pg.rs
@@ -1,9 +1,9 @@
-use chopin_pg::{PgConnection, PgConfig, PgResult};
+use chopin_pg::{PgConfig, PgConnection, PgResult};
 use std::time::Instant;
 
 fn main() -> PgResult<()> {
     let config = PgConfig::new("127.0.0.1", 5432, "chopin", "chopin", "chopin");
-    
+
     println!("Connecting to PostgreSQL...");
     let mut conn = match PgConnection::connect(&config) {
         Ok(c) => c,
@@ -19,28 +19,43 @@ fn main() -> PgResult<()> {
     // Simple SELECT 1 benchmark
     let iterations = 100_000;
     println!("Running {} iterations of 'SELECT 1'...", iterations);
-    
+
     let start = Instant::now();
     for _ in 0..iterations {
         let rows = conn.query("SELECT 1", &[])?;
         let _val: i32 = rows[0].get_i32(0)?.unwrap();
     }
     let duration = start.elapsed();
-    
-    println!("Throughput: {:.2} req/s", iterations as f64 / duration.as_secs_f64());
-    println!("Average latency: {:.2} µs", duration.as_micros() as f64 / iterations as f64);
+
+    println!(
+        "Throughput: {:.2} req/s",
+        iterations as f64 / duration.as_secs_f64()
+    );
+    println!(
+        "Average latency: {:.2} µs",
+        duration.as_micros() as f64 / iterations as f64
+    );
 
     // SQL with parameters
-    println!("\nRunning {} iterations of parameterized query...", iterations);
+    println!(
+        "\nRunning {} iterations of parameterized query...",
+        iterations
+    );
     let start = Instant::now();
     for i in 0..iterations {
-        let rows = conn.query("SELECT $1::int4 + $2::int4", &[&(i as i32), &10i32])?;
+        let rows = conn.query("SELECT $1::int4 + $2::int4", &[&i, &10i32])?;
         let _val: i32 = rows[0].get_i32(0)?.unwrap();
     }
     let duration = start.elapsed();
-    
-    println!("Throughput (parameterized): {:.2} req/s", iterations as f64 / duration.as_secs_f64());
-    println!("Average latency: {:.2} µs", duration.as_micros() as f64 / iterations as f64);
+
+    println!(
+        "Throughput (parameterized): {:.2} req/s",
+        iterations as f64 / duration.as_secs_f64()
+    );
+    println!(
+        "Average latency: {:.2} µs",
+        duration.as_micros() as f64 / iterations as f64
+    );
 
     // Clean up
     drop(conn);
